requirements.txt
certifi
chardet
idna
minimalmodbus
peewee
pyserial
requests
toml
urllib3
systemd-python
paho-mqtt